# [Release 6.4.1] Optimized wildcard emitter performance