--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 For changes before version 2.2.0, please see the commit history
 
+## [6.4.1] - 2020-05-10
+
+### Fixed
+- increased emitter performance in wildcard mode
+
 ## [6.4.0] - 2020-05-04
 
 ### Added