--- a/lib/eventemitter2.js
+++ b/lib/eventemitter2.js
@@ -416,12 +416,17 @@
   // It has zero elements if no any matches found and one or more
   // elements (leafs) if there are matches
   //
-  function searchListenerTree(handlers, type, tree, i) {
+  function searchListenerTree(handlers, type, tree, i, typeLength) {
     if (!tree) {
       return [];
     }
-    var listeners=[], leaf, len, branch, xTree, xxTree, isolatedBranch, endReached,
-        typeLength = type.length, currentType = type[i], nextType = type[i+1];
+    var listeners= null, branch, xTree, xxTree, isolatedBranch, endReached, currentType = type[i],
+        nextType = type[i + 1], branches, n, _listeners;
+
+    if (!typeLength) {
+      typeLength = type.length;
+    }
+
     if (i === typeLength && tree._listeners) {
       //
       // If at the end of the event(s) list and the tree has listeners
@@ -431,91 +436,117 @@
         handlers && handlers.push(tree._listeners);
         return [tree];
       } else {
-        for (leaf = 0, len = tree._listeners.length; leaf < len; leaf++) {
-          handlers && handlers.push(tree._listeners[leaf]);
-        }
+        handlers && handlers.push.apply(handlers, tree._listeners);
         return [tree];
       }
     }
 
-    if ((currentType === '*' || currentType === '**') || tree[currentType]) {
+    if (currentType === '*') {
       //
       // If the event emitted is '*' at this part
       // or there is a concrete match at this patch
       //
-      if (currentType === '*') {
-        for (branch in tree) {
-          if (branch !== '_listeners' && tree.hasOwnProperty(branch)) {
-            listeners = listeners.concat(searchListenerTree(handlers, type, tree[branch], i+1));
-          }
-        }
-        return listeners;
-      } else if(currentType === '**') {
-        endReached = (i+1 === typeLength || (i+2 === typeLength && nextType === '*'));
-        if(endReached && tree._listeners) {
-          // The next element has a _listeners, add it to the handlers.
-          listeners = listeners.concat(searchListenerTree(handlers, type, tree, typeLength));
-        }
-
-        for (branch in tree) {
-          if (branch !== '_listeners' && tree.hasOwnProperty(branch)) {
-            if(branch === '*' || branch === '**') {
-              if(tree[branch]._listeners && !endReached) {
-                listeners = listeners.concat(searchListenerTree(handlers, type, tree[branch], typeLength));
-              }
-              listeners = listeners.concat(searchListenerTree(handlers, type, tree[branch], i));
-            } else if(branch === nextType) {
-              listeners = listeners.concat(searchListenerTree(handlers, type, tree[branch], i+2));
-            } else {
-              // No match on this one, shift into the tree but not in the type array.
-              listeners = listeners.concat(searchListenerTree(handlers, type, tree[branch], i));
+      branches= ownKeys(tree);
+      n= branches.length;
+      while(n-->0){
+        branch= branches[n];
+        if (branch !== '_listeners') {
+          _listeners = searchListenerTree(handlers, type, tree[branch], i + 1, typeLength);
+          if(_listeners){
+            if(listeners){
+              listeners.push.apply(listeners, _listeners);
+            }else{
+              listeners = _listeners;
             }
           }
         }
-        return listeners;
+      }
+      return listeners;
+    } else if (currentType === '**') {
+      endReached = (i + 1 === typeLength || (i + 2 === typeLength && nextType === '*'));
+      if (endReached && tree._listeners) {
+        // The next element has a _listeners, add it to the handlers.
+        listeners = searchListenerTree(handlers, type, tree, typeLength, typeLength);
       }
 
-      listeners = listeners.concat(searchListenerTree(handlers, type, tree[currentType], i+1));
+      branches= ownKeys(tree);
+      n= branches.length;
+      while(n-->0){
+        branch= branches[n];
+        if (branch !== '_listeners') {
+          if (branch === '*' || branch === '**') {
+            if (tree[branch]._listeners && !endReached) {
+              _listeners = searchListenerTree(handlers, type, tree[branch], typeLength, typeLength);
+              if(_listeners){
+                if(listeners){
+                  listeners.push.apply(listeners, _listeners);
+                }else{
+                  listeners = _listeners;
+                }
+              }
+            }
+            _listeners = searchListenerTree(handlers, type, tree[branch], i, typeLength);
+          } else if (branch === nextType) {
+            _listeners = searchListenerTree(handlers, type, tree[branch], i + 2, typeLength);
+          } else {
+            // No match on this one, shift into the tree but not in the type array.
+            _listeners = searchListenerTree(handlers, type, tree[branch], i, typeLength);
+          }
+          if(_listeners){
+            if(listeners){
+              listeners.push.apply(listeners, _listeners);
+            }else{
+              listeners = _listeners;
+            }
+          }
+        }
+      }
+      return listeners;
+    }else if (tree[currentType]) {
+      listeners= searchListenerTree(handlers, type, tree[currentType], i + 1, typeLength);
     }
 
-    xTree = tree['*'];
+      xTree = tree['*'];
     if (xTree) {
       //
       // If the listener tree will allow any match for this part,
       // then recursively explore all branches of the tree
       //
-      searchListenerTree(handlers, type, xTree, i+1);
+      searchListenerTree(handlers, type, xTree, i + 1, typeLength);
     }
 
     xxTree = tree['**'];
-    if(xxTree) {
-      if(i < typeLength) {
-        if(xxTree._listeners) {
+    if (xxTree) {
+      if (i < typeLength) {
+        if (xxTree._listeners) {
           // If we have a listener on a '**', it will catch all, so add its handler.
-          searchListenerTree(handlers, type, xxTree, typeLength);
+          searchListenerTree(handlers, type, xxTree, typeLength, typeLength);
         }
 
         // Build arrays of matching next branches and others.
-        for(branch in xxTree) {
-          if(branch !== '_listeners' && xxTree.hasOwnProperty(branch)) {
-            if(branch === nextType) {
+        branches= ownKeys(xxTree);
+        n= branches.length;
+        while(n-->0){
+          branch= branches[n];
+          if (branch !== '_listeners') {
+            if (branch === nextType) {
               // We know the next element will match, so jump twice.
-              searchListenerTree(handlers, type, xxTree[branch], i+2);
-            } else if(branch === currentType) {
+              searchListenerTree(handlers, type, xxTree[branch], i + 2, typeLength);
+            } else if (branch === currentType) {
               // Current node matches, move into the tree.
-              searchListenerTree(handlers, type, xxTree[branch], i+1);
+              searchListenerTree(handlers, type, xxTree[branch], i + 1, typeLength);
             } else {
               isolatedBranch = {};
               isolatedBranch[branch] = xxTree[branch];
-              searchListenerTree(handlers, type, { '**': isolatedBranch }, i+1);
+              searchListenerTree(handlers, type, {'**': isolatedBranch}, i + 1, typeLength);
             }
           }
         }
-      } else if(xxTree._listeners) {
+      } else if (xxTree._listeners) {
         // We have reached the end and still on a '**'
-        searchListenerTree(handlers, type, xxTree, typeLength);
-      } else if(xxTree['*'] && xxTree['*']._listeners) {
-        searchListenerTree(handlers, type, xxTree['*'], typeLength);
+        searchListenerTree(handlers, type, xxTree, typeLength, typeLength);
+      } else if (xxTree['*'] && xxTree['*']._listeners) {
+        searchListenerTree(handlers, type, xxTree['*'], typeLength, typeLength);
       }
     }
 
@@ -844,6 +875,7 @@
 
     var type = arguments[0], ns, wildcard= this.wildcard, kind;
     var args,l,i,j, containsSymbol;
+    var delimiter= this.delimiter;
 
     if (type === 'newListener' && !this._newListener) {
       if (!this._events.newListener) {
@@ -854,24 +886,30 @@
     if (wildcard) {
       kind = typeof type;
       if (kind === 'string') {
-        ns = type.split(this.delimiter);
-      } else {
-        if(kind==='symbol'){
-          ns= [type];
-        }else{
-          ns = type.slice();
-          l= type.length;
-          if(symbolsSupported) {
-            for (i = 0; i < l; i++) {
-              if (typeof type[i] === 'symbol') {
-                containsSymbol = true;
-                break;
-              }
+        l = 0;
+        j = 0;
+        if ((i = type.indexOf(delimiter)) !== -1) {
+          ns = new Array(5);
+          do {
+            ns[l++] = type.slice(j, i);
+            j = i + 1;
+          } while ((i = type.indexOf(delimiter, j + 1)) !== -1);
+
+          ns[l++] = type.slice(j);
+        }
+      } else if (kind !== 'symbol') {
+        ns = type.slice();
+        l = type.length;
+        if (symbolsSupported) {
+          for (i = 0; i < l; i++) {
+            if (typeof type[i] === 'symbol') {
+              containsSymbol = true;
+              break;
             }
           }
-          if(!containsSymbol){
-            type = type.join(this.delimiter);
-          }
+        }
+        if (!containsSymbol) {
+          type = type.join(delimiter);
         }
       }
     }
@@ -902,7 +940,7 @@
 
     if (wildcard) {
       handler = [];
-      searchListenerTree.call(this, handler, ns, this.listenerTree, 0);
+      searchListenerTree.call(this, handler, ns || [type], this.listenerTree, 0, l);
     } else {
       handler = this._events[type];
       if (typeof handler === 'function') {
@@ -1190,8 +1228,8 @@
     if(this.wildcard) {
       var ns = typeof type === 'string' ? type.split(this.delimiter) : type.slice();
       leafs = searchListenerTree.call(this, null, ns, this.listenerTree, 0);
-    }
-    else {
+      if(!leafs) return this;
+    } else {
       // does not use listeners(), so no side effect of creating _events[type]
       if (!this._events[type]) return this;
       handlers = this._events[type];
@@ -1289,9 +1327,20 @@
     }
 
     if (this.wildcard) {
-      var ns = typeof type === 'string' ? type.split(this.delimiter) : type.slice();
-      var leafs = searchListenerTree.call(this, null, ns, this.listenerTree, 0);
+      var ns, i, l, j, delimiter= this.delimiter;
+      l = 0;
+      j = 0;
+      if ((i = type.indexOf(delimiter)) !== -1) {
+        ns = new Array(5);
+        do {
+          ns[l++] = type.slice(j, i);
+          j = i + 1;
+        } while ((i = type.indexOf(delimiter, j + 1)) !== -1);
 
+        ns[l++] = type.slice(j);
+      }
+      var leafs = searchListenerTree.call(this, null, ns || [type], this.listenerTree, 0, l);
+      if(!leafs) return this;
       for (var iLeaf = 0; iLeaf < leafs.length; iLeaf++) {
         var leaf = leafs[iLeaf];
         leaf._listeners = null;