--- a/lib/eventemitter2.js
+++ b/lib/eventemitter2.js
@@ -420,12 +420,35 @@
     if (!tree) {
       return [];
     }
+
+    if (i === 0) {
+      var kind = typeof type;
+      if (kind === 'string') {
+        var ns, n, l = 0, j = 0, delimiter = this.delimiter, dl = delimiter.length;
+        if ((n = type.indexOf(delimiter)) !== -1) {
+          ns = new Array(5);
+          do {
+            ns[l++] = type.slice(j, n);
+            j = n + dl;
+          } while ((n = type.indexOf(delimiter, j)) !== -1);
+
+          ns[l++] = type.slice(j);
+          type = ns;
+          typeLength = l;
+        } else {
+          type = [type];
+          typeLength = 1;
+        }
+      } else if (kind === 'object') {
+        typeLength = type.length;
+      } else {
+        type = [type];
+        typeLength = 1;
+      }
+    }
+
     var listeners= null, branch, xTree, xxTree, isolatedBranch, endReached, currentType = type[i],
         nextType = type[i + 1], branches, n, _listeners;
-
-    if (!typeLength) {
-      typeLength = type.length;
-    }
 
     if (i === typeLength && tree._listeners) {
       //
@@ -889,9 +912,8 @@
 
     this._events || init.call(this);
 
-    var type = arguments[0], ns, wildcard= this.wildcard, kind;
+    var type = arguments[0], ns, wildcard= this.wildcard;
     var args,l,i,j, containsSymbol;
-    var delimiter, dl;
 
     if (type === 'newListener' && !this._newListener) {
       if (!this._events.newListener) {
@@ -900,34 +922,21 @@
     }
 
     if (wildcard) {
-      delimiter= this.delimiter;
-      dl= delimiter.length;
-      kind = typeof type;
-      if (kind === 'string') {
-        l = 0;
-        j = 0;
-        if ((i = type.indexOf(delimiter)) !== -1) {
-          ns = new Array(5);
-          do {
-            ns[l++] = type.slice(j, i);
-            j = i + dl;
-          } while ((i = type.indexOf(delimiter, j)) !== -1);
-
-          ns[l++] = type.slice(j);
-        }
-      } else if (kind !== 'symbol') {
-        ns = type.slice();
-        l = type.length;
-        if (symbolsSupported) {
-          for (i = 0; i < l; i++) {
-            if (typeof type[i] === 'symbol') {
-              containsSymbol = true;
-              break;
+      ns= type;
+      if(type!=='newListener' && type!=='removeListener'){
+        if (typeof type === 'object') {
+          l = type.length;
+          if (symbolsSupported) {
+            for (i = 0; i < l; i++) {
+              if (typeof type[i] === 'symbol') {
+                containsSymbol = true;
+                break;
+              }
             }
           }
-        }
-        if (!containsSymbol) {
-          type = type.join(delimiter);
+          if (!containsSymbol) {
+            type = type.join(this.delimiter);
+          }
         }
       }
     }
@@ -958,7 +967,7 @@
 
     if (wildcard) {
       handler = [];
-      searchListenerTree.call(this, handler, ns || [type], this.listenerTree, 0, l);
+      searchListenerTree.call(this, handler, ns, this.listenerTree, 0, l);
     } else {
       handler = this._events[type];
       if (typeof handler === 'function') {
@@ -1026,7 +1035,7 @@
 
     this._events || init.call(this);
 
-    var type = arguments[0], wildcard= this.wildcard, ns, kind, containsSymbol;
+    var type = arguments[0], wildcard= this.wildcard, ns, containsSymbol;
     var args,l,i,j;
 
     if (type === 'newListener' && !this._newListener) {
@@ -1034,16 +1043,11 @@
     }
 
     if (wildcard) {
-      kind = typeof type;
-      if (kind === 'string') {
-        ns = type.split(this.delimiter);
-      } else {
-        if(kind==='symbol'){
-          ns= [type];
-        }else{
-          ns = type.slice();
-          l= type.length;
-          if(symbolsSupported) {
+      ns= type;
+      if(type!=='newListener' && type!=='removeListener'){
+        if (typeof type === 'object') {
+          l = type.length;
+          if (symbolsSupported) {
             for (i = 0; i < l; i++) {
               if (typeof type[i] === 'symbol') {
                 containsSymbol = true;
@@ -1051,7 +1055,7 @@
               }
             }
           }
-          if(!containsSymbol){
+          if (!containsSymbol) {
             type = type.join(this.delimiter);
           }
         }
@@ -1345,22 +1349,10 @@
     }
 
     if (this.wildcard) {
-      var ns, i, l, j, delimiter= this.delimiter, dl= delimiter.length;
-      l = 0;
-      j = 0;
-      if ((i = type.indexOf(delimiter)) !== -1) {
-        ns = new Array(5);
-        do {
-          ns[l++] = type.slice(j, i);
-          j = i + dl;
-        } while ((i = type.indexOf(delimiter, j)) !== -1);
-
-        ns[l++] = type.slice(j);
-      }
-      var leafs = searchListenerTree.call(this, null, ns || [type], this.listenerTree, 0, l);
-      if(!leafs) return this;
-      for (var iLeaf = 0; iLeaf < leafs.length; iLeaf++) {
-        var leaf = leafs[iLeaf];
+      var leafs = searchListenerTree.call(this, null, type, this.listenerTree, 0), leaf, i;
+      if (!leafs) return this;
+      for (i = 0; i < leafs.length; i++) {
+        leaf = leafs[i];
         leaf._listeners = null;
       }
       this.listenerTree && recursivelyGarbageCollect(this.listenerTree);