--- a/lib/eventemitter2.js
+++ b/lib/eventemitter2.js
@@ -418,7 +418,7 @@
   //
   function searchListenerTree(handlers, type, tree, i, typeLength) {
     if (!tree) {
-      return [];
+      return null;
     }
 
     if (i === 0) {