--- a/lib/eventemitter2.js
+++ b/lib/eventemitter2.js
@@ -448,7 +448,7 @@
     }
 
     var listeners= null, branch, xTree, xxTree, isolatedBranch, endReached, currentType = type[i],
-        nextType = type[i + 1], branches, n, _listeners;
+        nextType = type[i + 1], branches, _listeners;
 
     if (i === typeLength && tree._listeners) {
       //