@DigitalBrainJS 2020-05-11
> EventEmitterHeatUp x 3,397,135 ops/sec ±3.47% (71 runs sampled)
> EventEmitter x 3,728,915 ops/sec ±3.55% (74 runs sampled)
> EventEmitter2 x 12,067,339 ops/sec ±4.83% (67 runs sampled)
> **EventEmitter2 (wild) x 277,835 ops/sec ±3.57% (68 runs sampled)**
> EventEmitter2 (wild) using plain events x 382,594 ops/sec ±3.60% (67 runs sampled)
> **EventEmitter2 (wild) emitting ns x 1,725,874 ops/sec ±4.97% (66 runs sampled)
> EventEmitter2 (wild) emitting a plain event x 2,903,992 ops/sec ±4.31% (71 runs sampled)**
> EventEmitter3 x 10,763,520 ops/sec ±3.86% (65 runs sampled)
> 
> Fastest is EventEmitter2