@DigitalBrainJS 2020-05-19
@RangerMauve It would be nice if you could complete the last step for CI integration- setup the coveralls.io service. You just need to get an environment variable for the Travis build :)