@RangerMauve 2020-05-19
@DigitalBrainJS Yeah sure, where do I get the veriable and where should I put it? 😁