@DigitalBrainJS 2020-05-19
@RangerMauve https://github.com/EventEmitter2/EventEmitter2/issues/249#issuecomment-625231761